+++ Documents/Help/Changelog.md
@@ -6,7 +6,7 @@
 
 ## 2026-11-xx - Build 2611 - November 2026
 
-* Implemented [#511](https://github.com/Krypton-Suite/Extended-Toolkit/issues/511), `ExtendedKryptonMessageBox` Expandable Footer Feature
+* Implemented [#511](https://github.com/Krypton-Suite/Extended-Toolkit/issues/511), `KryptonMessageBoxExtended` Expandable Footer Feature
   - **New Expandable Footer** - Similar to Windows TaskDialog, the message box now supports an expandable footer area
   - **Collapsed/Expanded States** - Footer can start collapsed or expanded, with user toggle capability
   - **Toggle Link** - "Show details" / "Hide details" link allows users to expand/collapse the footer